Move normalize_grayscale.py into the Python/Filters directory
